Move modal dialog components to dialog directory
