+++ ui/src/main/webapp/src/app/app.module.ts
@@ -27,7 +27,7 @@ import {ConfigurationComponent} from "./configuration/configuration.component";
 import {TechnologyComponent} from "./configuration/technology.component";
 import {RulesModalComponent} from "./configuration/rules-modal.component";
 import {AddRulesPathModalComponent} from "./configuration/add-rules-path-modal.component";
-import {ConfirmationModalComponent} from "./shared/confirmation-modal.component";
+import {ConfirmationModalComponent} from "./shared/dialog/confirmation-modal.component";
 import {CustomRuleSelectionComponent} from "./analysis-context/custom-rule-selection.component";
 
 import {KeycloakService} from "./core/authentication/keycloak.service";
@@ -37,7 +37,7 @@ import {UploadQueueComponent} from "./shared/upload/upload-queue.component";
 import {UploadProgressbarComponent} from "./shared/upload/upload-progressbar.component";
 import {AnalysisContextAdvancedOptionsModalComponent} from "./analysis-context/analysis-context-advanced-options-modal.component";
 import {ConfigurationOptionsService} from "./configuration/configuration-options.service";
-import {ModalDialogComponent} from "./shared/modal-dialog.component";
+import {ModalDialogComponent} from "./shared/dialog/modal-dialog.component";
 import {NotificationService} from "./core/notification/notification.service";
 import {NotificationComponent} from "./shared/notification.component";
 import {ConfirmDeactivateGuard} from "./shared/confirm-deactivate.guard";
