+++ ui/src/main/webapp/src/app/shared/dialog/confirmation-modal.component.ts
